+++ src/App.jsx
@@ -20,4 +20,4 @@ const App = () => (
         </div>
     </div>
 );
-export default App
+export default App;
