+++ src/components/content/ContentItem.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 
 const ContentItem = ({img, description, price}) => (
-  <div className="product-item">
-    <div className="img-main-photo">
-      <img src={img} alt={img}/>
+    <div className="product-item">
+        <div className="img-main-photo">
+            <img src={img} alt=""/>
+        </div>
+        <div className="product-description">
+            <p>{description}</p>
+            <p className="price">{price}</p>
+        </div>
     </div>
-    <div className="product-description">
-      <p>{description}</p>
-      <p className="price">{price}</p>
-    </div>
-  </div>
 );
 
 export default ContentItem;
