+++ src/components/footer/index.js
@@ -1,3 +1,3 @@
-import Footer from "./footer";
+import Footer from "./Footer";
 
 export default Footer;
