+++ src/components/header/index.js
@@ -1,3 +1,3 @@
-import Header from "./header";
+import Header from "./Header";
 
 export default Header;
