+++ src/components/sidebar/index.js
@@ -1,3 +1,3 @@
-import Sidebar from "./sidebar";
+import Sidebar from "./Sidebar";
 
 export default Sidebar;
