+++ src/components/sidebar/sidebar.css
@@ -6,6 +6,8 @@
     display: flex;
     justify-content: space-between;
     border-bottom: 2px solid rgba(0, 0, 0, .1);
+    margin: 10px 0;
+    padding: 0 0 5px 0;
 }
 
 .sidebar .sortFilter p {
