+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './common.css';
-import App from './app';
+import App from './App';
 
 ReactDOM.render(<App/>, document.getElementById('root'));
